ceas/pyproject.toml
[project]
name = "ceas"
version = "0.1.0"
description = ""
authors = [
    {name = "devgiordane",email = "bruno_giordane@outlook.com"}
]
readme = "README.md"
requires-python = "~3.13"
dependencies = [
    "apscheduler",
    "requests",
    "pyyaml",
    "click",
]

[tool.poetry]
packages = [{include = "ceas"}]

[tool.poetry.scripts]
ceas = "ceas.cli:cli"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
